Split AdvancedBusExtensions to multiple files
